refactor(guard): extends role permission
