+++ libs/common/src/guards/index.ts
@@ -1,3 +1,4 @@
 export * from './auth.guard';
 export * from './admin.guard';
 export * from './mod.guard';
+export * from './super-admin.guard';
