+++ libs/common/src/guards/super-admin.guard.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@nestjs/common';
+import { AuthCoreGuard } from './auth.core.guard';
+import { UserRole } from '@app/resource/users/enums';
+
+@Injectable()
+export class SuperAdminGuard extends AuthCoreGuard {
+    constructor() {
+        super();
+        this._acceptRoles.push(UserRole.SuperAdmin);
+    }
+}
